Create file to get input data from Advent of Code
This means that I don't have to upload my own inputs to GitHub, and
should make it somewhat (?) easier for other people to use my solutions.

Actually, copying and pasting a bunch of text from the web into a file
is easier than finding an obscure cookie and setting it as an
environment variable... Regardless, this still feels better to me.